pc1 planchas

+++ 3er ciclo/planchas pc1/funciones.h
@@ -1,7 +1,7 @@
 #ifndef __FUNCIONES_H__
 #define __FUNCIONES_H__
 
-bool natural(double n);
+bool natural(int n);
 double calcularXrecursiva(int i, int n);
 
 #endif // __FUNCIONES_H__
